RavenDB-21129 Renaming
