+++ src/Raven.Studio/typescript/components/pages/resources/manageServer/serverWideAnalyzers/CustomServerWideAnalyzers.stories.tsx
@@ -1,7 +1,7 @@
 ﻿import React from "react";
 import { Meta, StoryObj } from "@storybook/react";
 import { withStorybookContexts, withBootstrap5 } from "test/storybookTestUtils";
-import ServerWideAnalyzers from "./ServerWideAnalyzers";
+import ServerWideAnalyzers from "./CustomServerWideAnalyzers";
 import { DatabasesStubs } from "test/stubs/DatabasesStubs";
 import { mockServices } from "test/mocks/services/MockServices";
