+++ src/Raven.Studio/typescript/components/pages/resources/manageServer/serverWideAnalyzers/CustomServerWideAnalyzers.tsx
